+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 - 
 
 ### Changed
-- 
+- Make calibration work, maybe with interp1d
+-
 
 ### Removed
 -